pyproject.toml
[tool.poetry]
name = "lockerr"
version = "0.1.0"
description = "a Discord bot for hindering users from switching voice channels"
authors = ["Quoorex <46283604+Quoorex@users.noreply.github.com>"]
license = "Apache-2.0"

[tool.poetry.dependencies]
python = "^3.6"
"discord.py" = "^1.6.0"
python-dotenv = "^0.15.0"

[tool.poetry.dev-dependencies]
black = "^20.8b1"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
